requirements.txt
# Core Django dependencies
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0
django-guardian>=2.4.0
django-extensions>=3.2.0

# Database
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0

# Authentication & Security
django-allauth>=0.57.0
djangorestframework-simplejwt>=5.2.0
django-crispy-forms>=2.0.0
crispy-bootstrap5>=0.7.0
django-widget-tweaks>=1.5.0

# API Documentation
drf-spectacular>=0.27.0
drf-spectacular[sidecar]

# File handling & Images
Pillow>=10.0.0
django-storages>=1.14.0

# Data processing & Export
pandas>=2.0.0
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Utilities
python-decouple>=3.8
celery>=5.3.0
redis>=4.5.0
python-dateutil>=2.8.0
pytz>=2023.3

# Development & Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
model-bakery>=1.17.0
django-debug-toolbar>=4.2.0

# Linting & Formatting
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
pre-commit>=3.3.0

# Production
gunicorn>=21.0.0
whitenoise>=6.5.0
sentry-sdk>=1.32.0

# Phone validation and PDF generation
phonenumbers>=8.13.0
reportlab>=4.0.0

# HTMX for frontend interactivity
django-htmx>=1.16.0
